// packages/shared/src/modules/exercises/enums.ts
export enum Muscle {
  Abdominals = 'abdominals',
  Hamstrings = 'hamstrings',
  Adductors = 'adductors',
  Quadriceps = 'quadriceps',
  Biceps = 'biceps',
  Shoulders = 'shoulders',
  Chest = 'chest',
  MiddleBack = 'middle back',
  Calves = 'calves',
  Glutes = 'glutes',
  LowerBack = 'lower back',
  Lats = 'lats',
  Triceps = 'triceps',
  Traps = 'traps',
  Forearms = 'forearms',
  Neck = 'neck',
  Abductors = 'abductors',
}

export enum ForceType {
  Pull = 'pull',
  Push = 'push',
  Static = 'static',
}

export enum LevelType {
  Beginner = 'beginner',
  Intermediate = 'intermediate',
  Expert = 'expert',
}

export enum MechanicType {
  Compound = 'compound',
  Isolation = 'isolation',
}

export enum EquipmentType {
  BodyOnly = 'body only',
  Machine = 'machine',
  Other = 'other',
  FoamRoll = 'foam roll',
  Kettlebells = 'kettlebells',
  Dumbbell = 'dumbbell',
  Cable = 'cable',
  Barbell = 'barbell',
  MedicineBall = 'medicine ball',
  Bands = 'bands',
  ExerciseBall = 'exercise ball',
  EZCurlBar = 'e-z curl bar',
}

export enum CategoryType {
  Strength = 'strength',
  Stretching = 'stretching',
  Plyometrics = 'plyometrics',
  Strongman = 'strongman',
  Powerlifting = 'powerlifting',
  Cardio = 'cardio',
  OlympicWeightlifting = 'olympic weightlifting',
}